refactor(minifyConditionalComments): rewrite typescript
