+++ src/index.ts
@@ -65,7 +65,7 @@ const modules = {
     // example: () => import('./_modules/example.mjs'),
     mergeScripts: () => import('./_modules/mergeScripts.mjs'),
     mergeStyles: () => import('./_modules/mergeStyles.mjs'),
-    minifyConditionalComments: () => import('./_modules/minifyConditionalComments.mjs'),
+    minifyConditionalComments: () => interop(import('./_modules/minifyConditionalComments')),
     minifyCss: () => import('./_modules/minifyCss.mjs'),
     minifyJs: () => import('./_modules/minifyJs.mjs'),
     minifyJson: () => interop(import('./_modules/minifyJson.js')),
